Added dynamic data for all the pages
