+++ src/pages/_app.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import Layout from "../components/templates/Layout";
+import Layout from "@components/templates/Layout";
 import type { AppProps } from "next/app";
 import type { NextPage } from "next";
-import "../styles/global.scss";
+import "@styles/global.scss";
 
 type NextPageWithLayout = NextPage & {
-  getTitle?: (pageProps: any) => string;
+  getTitle?: (title: string) => string;
 };
 
 type AppPropsWithLayout = AppProps & {
@@ -15,10 +15,13 @@ type AppPropsWithLayout = AppProps & {
 const app = ({ Component, pageProps }: AppPropsWithLayout) => {
   const getTitle = Component?.getTitle ?? ((title) => title);
   const title = getTitle(pageProps);
-
+  
+  const {props} = pageProps;
+  console.log('pageProps', pageProps, props)
   return (
     <Layout title={title}>
-      <Component {...pageProps} />
+      {props && <Component {...props} />}
+      {!props && <h1>no data</h1>}
     </Layout>
   );
 };
