+++ src/pages/about.tsx
@@ -1,20 +1,21 @@
 import About from "@components/templates/About";
-import profilePic from "../../public/me.jpg";
+import { ApiNameEnum, IAbout } from "@interfaces/index";
+import { fetchData } from "@utils/fetch";
 
-const AboutPage = () => {
-  const heading = "Hey!👋 I'm Saurabh";
-
-  const about = [
-    "I'm a Senior Software Enginner at Nagarro, originally from Varanasi, India.",
-    "A confident, professional, creative and resourceful person with the ability to work efficiently in fast-paced, multitasking environments, along with maintaining a positive attitude and strong work ethics.",
-    "My areas of interest lie in Frontend Development, Web developement, UI automation and DevOps. I have experience building enterprise grade WebApps and scalable frontend architectures.",
-    "I also enjoy playing the table-tennis, vollyball, listening songs and playing video games.",
-  ];
-  return <About heading={heading} about={about} profilePic={profilePic} />;
+const AboutPage = (props: IAbout) => {
+  return <About {...props} />;
 };
 
 AboutPage.getTitle = () => {
   return "About";
 };
+
+export async function getStaticProps() {
+  const data = await fetchData(ApiNameEnum.ABOUT);
+
+  return {
+    props: { props: data },
+  };
+}
 
 export default AboutPage;
