+++ src/pages/api/[...slug].ts
@@ -1,22 +1,30 @@
-import { NextApiRequest, NextApiResponse } from 'next'
-import { homeData } from '@utils/home-data';
-import { ApiNameEnum } from '@interfaces/index';
+import { NextApiRequest, NextApiResponse } from "next";
+import { homeData, aboutData, skillsData, experienceData } from "@utils/site-data";
+import { ApiNameEnum } from "@interfaces/index";
 
 const handler = (request: NextApiRequest, response: NextApiResponse) => {
   try {
-    const {slug} = request.query;
+    const { slug } = request.query;
     const res = getData(slug[0]);
-      response.status(200).json(res);
+    response.status(200).json(res);
   } catch (err: any) {
-    response.status(500).json({ statusCode: 500, message: err.message })
+    response.status(500).json({ statusCode: 500, message: err.message });
   }
-}
+};
 
-const getData =(slug: string | ApiNameEnum) =>{
-  switch(slug){
+const getData = (slug: string | ApiNameEnum) => {
+  switch (slug) {
     case ApiNameEnum.HOME:
       return homeData;
+    case ApiNameEnum.ABOUT:
+      return aboutData;
+    case ApiNameEnum.SKILLS:
+      return skillsData;
+    case ApiNameEnum.EXPERIENCES:
+      return experienceData;
+    default:
+      return null;
   }
-}
+};
 
-export default handler
+export default handler;
