--- src/utils/home-data.ts
@@ -1,7 +1,0 @@
-import { IHome } from "@interfaces/index";
-
-export const homeData: IHome = {
-  name: "SAURABH GHOSH",
-  roles: ["a Javascript Developer", "a Blogger"],
-  data: "I'm",
-}
